requirements.txt
Owlready2==0.24
rdflib==5.0.0
python-slugify>=4.0.1
